Validação do anunciar

+++ comandos/anunciar.js
@@ -1,4 +1,6 @@
 module.exports.run = (client, message, args) => {
+    if (!args[0]) return message.reply("Você precisa informar o texto a ser anunciado após o comando.")
+    
     message.delete();
 
     let messageSend = args.join(' ')
